user describe.each and convert integ tests to unit tests
